+++ database/migrations/2020_10_03_175504_create_bcps_table.php
@@ -14,11 +14,11 @@ class CreateBcpsTable extends Migration
     public function up()
     {
         Schema::create('bcps', function (Blueprint $table) {
-            $table->bigIncrements('id');
+            $table->increments('id');
             $table->string('executive_summary');
             $table->text('introduction');
             $table->text('planning_assumptions');
-            $table->unsignedBigInteger('wsp_id')->nullable();
+            $table->unsignedInteger('wsp_id')->nullable();
             $table->foreign('wsp_id')->references('id')->on('wsps');
             $table->timestamp('created_at')->nullable();
             $table->timestamp('updated_at')->nullable();
