+++ database/migrations/2020_10_03_175512_create_revenues_table.php
@@ -14,14 +14,15 @@ class CreateRevenuesTable extends Migration
     public function up()
     {
         Schema::create('revenues', function (Blueprint $table) {
-            $table->bigIncrements('id');
+            $table->increments('id');
             $table->string('amount');
             $table->string('month');
             $table->string('year');
-            $table->unsignedBigInteger('wsp_id')->nullable();
+            $table->unsignedInteger('wsp_id')->nullable();
             $table->foreign('wsp_id')->references('id')->on('wsps');
             $table->timestamp('created_at')->nullable();
-            $table->timestamp('updated_at')->nullable();
+            $table->timestamp('updated_at')->nullable();
+
         });
     }
