+++ database/migrations/2020_10_03_175516_create_eoi_estimatedcost_table.php
@@ -14,15 +14,16 @@ class CreateEoiEstimatedcostTable extends Migration
     public function up()
     {
         Schema::create('eoi_estimatedcost', function (Blueprint $table) {
-            $table->bigIncrements('id');
+            $table->increments('id');
             $table->string('unit');
             $table->string('total');
             $table->timestamp('created_at')->nullable();
             $table->timestamp('updated_at')->nullable();
-            $table->unsignedBigInteger('eoi_id')->nullable();
+            $table->unsignedInteger('eoi_id')->nullable();
             $table->foreign('eoi_id')->references('id')->on('eois');
-            $table->unsignedBigInteger('estimatedcost_id')->nullable();
-            $table->foreign('estimatedcost_id')->references('id')->on('estimatedcosts');
+            $table->unsignedInteger('estimatedcost_id')->nullable();
+            $table->foreign('estimatedcost_id')->references('id')->on('estimatedcosts');
+
         });
     }
