+++ database/migrations/2020_10_03_175517_create_connection_eoi_table.php
@@ -14,15 +14,16 @@ class CreateConnectionEoiTable extends Migration
     public function up()
     {
         Schema::create('connection_eoi', function (Blueprint $table) {
-            $table->bigIncrements('id');
+            $table->increments('id');
             $table->string('areas');
             $table->integer('total');
             $table->timestamp('created_at')->nullable();
             $table->timestamp('updated_at')->nullable();
-            $table->unsignedBigInteger('connection_id')->nullable();
+            $table->unsignedInteger('connection_id')->nullable();
             $table->foreign('connection_id')->references('id')->on('connections');
-            $table->unsignedBigInteger('eoi_id')->nullable();
-            $table->foreign('eoi_id')->references('id')->on('eois');
+            $table->unsignedInteger('eoi_id')->nullable();
+            $table->foreign('eoi_id')->references('id')->on('eois');
+
         });
     }
